Configure my.cnf default values
character-set-server is set to utf8, default is latin1 [1]
collation-server is set to utf8_general_ci, default is latin1_swedish_ci

innodb_buffer_pool_size is set to 512M, the default value is 128M.
innodb_file_format is set to Barracuda, the default value is Antelope.
innodb_file_per_table is set to 1, the default value is 0.
innodb_flush_log_at_trx_commit is set to 2, the default value is 1.
innodb_flush_method is set to O_DIRECT, the deefault value is NULL
innodb_io_capacity is set to 500, the default vaule is set to 200
innodb_read_io_threads is set to 8, the default vaue is 4
innodb_write_io_threads is set to 8, the default value is 4

open_files_limit is set to 102400, the default value is 5000 [3]

[1] https://dev.mysql.com/doc/refman/5.6/en/server-options.html#option_mysqld_character-set-server
[2] https://dev.mysql.com/doc/refman/5.6/en/innodb-parameters.html
[3] https://dev.mysql.com/doc/refman/5.6/en/server-system-variables.html

Change-Id: I5bec63612d9230ab1a69c5d32528b0021473dff1